Add new post submission exception
