+++ app/authentication/exceptions.py
@@ -1,0 +1,16 @@
+class NoTokenException(Exception):
+    def __init__(self, value):
+        super().__init__()
+        self.value = value
+
+    def __str__(self):
+        return self.value
+
+
+class PostSubmissionException(Exception):
+    def __init__(self, value):
+        super().__init__()
+        self.value = value
+
+    def __str__(self):
+        return self.value
