--- app/authentication/no_token_exception.py
@@ -1,7 +1,0 @@
-class NoTokenException(Exception):
-    def __init__(self, value):
-        super().__init__()
-        self.value = value
-
-    def __str__(self):
-        return self.value
