+++ templates/errors/submission-complete.html
@@ -1,0 +1,8 @@
+{% extends 'layouts/_base.html' %}
+
+{% set page_title = _("No token") ~ " - " ~ _("Census 2021") %}
+
+{% block main %}
+<h1>{{ _("This page is no longer available") }}</h1>
+<p>{{ _("Your census has been submitted.") }}</p>
+{% endblock %}
